refactor: split test files to comply with 1-array-per-file rule
Split props.test.tsx into:
- title-message.test.tsx (individual tests)
- confirm-button-labels.test.tsx (array-based)
- cancel-button-labels.test.tsx (array-based)

Split accessibility.test.tsx into:
- aria-attributes.test.tsx (array-based)
- dynamic-id.test.tsx (individual tests)

This complies with array-based-test.md rule: "テストをまとめた配列は、
1ファイルに1つとし、それ以外のテストケースは別ファイルに切り出す"
